server/requirements.txt
pydantic
fastapi
requests
newspaper4k
lxml[html_clean]
openai
psutil
python-dotenv
uvicorn